Added sections to settings screen.
