+++ lib/src/data/models/article.dart
@@ -1,0 +1,13 @@
+class Article {
+  final String title;
+  final String imageUrl;
+  final String author;
+  final String postedOn;
+
+  Article({
+    required this.title,
+    required this.imageUrl,
+    required this.author,
+    required this.postedOn,
+  });
+}
