+++ lib/src/ui/screens/settings_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:tasker/src/data/providers/theme_provider.dart';
+import 'package:tasker/src/ui/widgets/single_section.dart';
 
 class SettingsScreen extends StatelessWidget {
   const SettingsScreen({super.key});
@@ -8,12 +9,24 @@ class SettingsScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SafeArea(
-      child: SwitchListTile(
-        title: Text('Dark Mode'),
-        value: Provider.of<ThemeProvider>(context).isDarkMode,
-        onChanged: (value) {
-          Provider.of<ThemeProvider>(context, listen: false).toggleTheme();
-        },
+      child: ListView(
+        children: [
+          SingleSection(
+            title: "Settings",
+            children: [
+              SwitchListTile(
+                title: Text('Dark Mode'),
+                value: Provider.of<ThemeProvider>(context).isDarkMode,
+                onChanged: (value) {
+                  Provider.of<ThemeProvider>(
+                    context,
+                    listen: false,
+                  ).toggleTheme();
+                },
+              ),
+            ],
+          ),
+        ],
       ),
     );
   }
