+++ lib/src/ui/widgets/article_card.dart
@@ -1,0 +1,74 @@
+import 'package:flutter/material.dart';
+import 'package:tasker/src/data/models/article.dart';
+
+class ArticleCard extends StatelessWidget {
+  const ArticleCard({super.key, required this.article});
+
+  final Article article;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 136,
+      margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8.0),
+      decoration: BoxDecoration(
+        border: Border.all(color: const Color(0xFFE0E0E0)),
+        borderRadius: BorderRadius.circular(8.0),
+      ),
+      padding: const EdgeInsets.all(8),
+      child: Row(
+        children: [
+          Expanded(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  article.title,
+                  style: const TextStyle(fontWeight: FontWeight.bold),
+                  maxLines: 2,
+                  overflow: TextOverflow.ellipsis,
+                ),
+                const SizedBox(height: 8),
+                Text(
+                  "${article.author} · ${article.postedOn}",
+                  style: Theme.of(context).textTheme.bodyMedium,
+                ),
+                const SizedBox(height: 8),
+                Row(
+                  mainAxisSize: MainAxisSize.min,
+                  children:
+                      [
+                        Icons.bookmark_border_rounded,
+                        Icons.share,
+                        Icons.more_vert,
+                      ].map((e) {
+                        return InkWell(
+                          onTap: () {},
+                          child: Padding(
+                            padding: const EdgeInsets.only(right: 8.0),
+                            child: Icon(e, size: 16),
+                          ),
+                        );
+                      }).toList(),
+                ),
+              ],
+            ),
+          ),
+          Container(
+            width: 100,
+            height: 100,
+            decoration: BoxDecoration(
+              color: Colors.grey,
+              borderRadius: BorderRadius.circular(8.0),
+              image: DecorationImage(
+                fit: BoxFit.cover,
+                image: NetworkImage(article.imageUrl),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
