+++ lib/src/ui/widgets/single_section.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+class SingleSection extends StatelessWidget {
+  const SingleSection({super.key, required this.title, required this.children});
+
+  final String title;
+  final List<Widget> children;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.start,
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        const SizedBox(height: 16),
+        Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Text(
+            title,
+            style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),
+          ),
+        ),
+        SizedBox(width: double.infinity, child: Column(children: children)),
+      ],
+    );
+  }
+}
